Refactor: Migrated from Lucia Auth to self rolled auth
